Added compatibility with relative line references (e.g. line=+12) in <location> tags.

+++ src/internal/XlsxParser.cpp
@@ -64,7 +64,7 @@ auto XlsxParser::parse() const -> Result
             }
             auto list = loc.split(QStringLiteral(" - "));
             msg.locations.emplace_back(
-                std::make_pair(list.first(), list.last().toInt()));
+                std::make_pair(list.first(), list.last()));
         }
 
         auto it =
